kvaser-bridge/requirements.txt
python-can==4.4.2
websockets>=12.0
pyinstaller>=6.0.0
